+++ src/fhir/R4R5-common/sdc-export.js
@@ -14,16 +14,8 @@
  * convertLFormsToQuestionnaireResponse()
  * -- Generate FHIR (standard or SDC) QuestionnaireResponse data from captured data in LForms
  */
-var sdcVersion = '2.7';
-var fhirVersionNum = '4.0';
 
 var self = {
-
-  SDCVersion: sdcVersion,
-  QProfile: 'http://hl7.org/fhir/uv/sdc/StructureDefinition/sdc-questionnaire|'+sdcVersion,
-  QRProfile: 'http://hl7.org/fhir/uv/sdc/StructureDefinition/sdc-questionnaireresponse|'+sdcVersion,
-  stdQProfile: 'http://hl7.org/fhir/'+fhirVersionNum+'/StructureDefinition/Questionnaire',
-  stdQRProfile: 'http://hl7.org/fhir/'+fhirVersionNum+'/StructureDefinition/QuestionnaireResponse',
 
   /**
    *  Convert LForms captured data to a bundle consisting of a FHIR SDC
@@ -32,8 +24,6 @@ var self = {
    *  extension).
    *
    * @param lfData a LForms form object
-   * @param noExtensions a flag that a standard FHIR Questionnaire is to be created without any extensions.
-   *  The default is false.
    * @param subject A local FHIR resource that is the subject of the output resource.
    *  If provided, a reference to this resource will be added to the output FHIR
    *  resource when applicable.
@@ -43,8 +33,8 @@ var self = {
    *  wish to put all of the returned resources into a transaction Bundle for
    *  creating them on a FHIR server.
    */
-  convertLFormsToQRAndExtracFHIRData: function(lfData, noExtensions, subject) {
-    var qr = this.convertLFormsToQuestionnaireResponse(lfData, noExtensions, subject);
+  convertLFormsToQRAndExtracFHIRData: function(lfData, subject) {
+    var qr = this.convertLFormsToQuestionnaireResponse(lfData, subject);
     if (!qr.id) {
       qr.id = this._commonExport._getUniqueId(
         qr.identifier && qr.identifier.value || 'QR')
@@ -208,10 +198,8 @@ var self = {
    *  Processes settings for a list field with choices.
    * @param targetItem an item in FHIR SDC Questionnaire object
    * @param item an item in the LForms form object
-   * @param noExtensions a flag that a standard FHIR Questionnaire is to be created without any extensions.
-   *        The default is false.
    */
-  _handleChoiceField: function(targetItem, item, noExtensions) {
+  _handleChoiceField: function(targetItem, item) {
     // an extension for the search url of the auto-complete field.
     if(item.externallyDefined) {
       this._handleExternallyDefined(targetItem, item);
@@ -220,7 +208,7 @@ var self = {
     else if (item.answers && !item.answerValueSet) {
       // Make sure the answers did not come from answerExpression.
       if (!item._fhirExt || !item._fhirExt[this.fhirExtAnswerExp])
-        targetItem.answerOption = this._handleAnswers(item, noExtensions);
+        targetItem.answerOption = this._handleAnswers(item);
     }
     else if (item.answerValueSet)
       targetItem.answerValueSet = item.answerValueSet;
@@ -230,12 +218,10 @@ var self = {
   /**
    * Process an item's answer list
    * @param item an item in the LForms form object
-   * @param noExtensions a flag that a standard FHIR Questionnaire is to be created without any extensions.
-   *        The default is false.
    * @returns {Array}
    * @private
    */
-  _handleAnswers: function(item, noExtensions) {
+  _handleAnswers: function(item) {
     var optionArray = [];
     for (var i=0, iLen=item.answers.length; i<iLen; i++) {
       var answer = item.answers[i];
@@ -274,25 +260,23 @@ var self = {
         option[valueKey] = answer.text;
       }
 
-      // needs an extension for label
-      if (!noExtensions) {
-        var ext = [];
-        if(answer.label) {
-          ext.push({
-            "url" : "http://hl7.org/fhir/StructureDefinition/questionnaire-optionPrefix",
-            "valueString" : answer.label
-          });
-        }
+      // label
+      var ext = [];
+      if(answer.label) {
+        ext.push({
+          "url" : "http://hl7.org/fhir/StructureDefinition/questionnaire-optionPrefix",
+          "valueString" : answer.label
+        });
+      }
 
-        if (answer.score !== null && answer.score !== undefined) {
-          ext.push({
-            "url" : "http://hl7.org/fhir/StructureDefinition/ordinalValue",
-            "valueDecimal" : parseFloat(answer.score)
-          });
-        }
-        if(ext.length > 0) {
-          option.extension = ext;
-        }
+      if (answer.score !== null && answer.score !== undefined) {
+        ext.push({
+          "url" : "http://hl7.org/fhir/StructureDefinition/ordinalValue",
+          "valueDecimal" : parseFloat(answer.score)
+        });
+      }
+      if(ext.length > 0) {
+        option.extension = ext;
       }
 
       optionArray.push(option);