+++ src/fhir/common/sdc-import-common.js
@@ -1337,6 +1337,7 @@ function addCommonSDCImportFns(ns) {
         if (answers) {
           item.answers = answers;
           lfData._updateAutocompOptions(item);
+          lfData._resetItemValueWithAnswers(item);
         }
         else { // if not already loaded
           if (expURL) {
@@ -1355,6 +1356,7 @@ function addCommonSDCImportFns(ns) {
                   LForms._valueSetAnswerCache[expURL] = answers;
                   item.answers = answers;
                   lfData._updateAutocompOptions(item);
+                  lfData._resetItemValueWithAnswers(item);
                 }
               }
             }).catch(function(error) {
@@ -1372,6 +1374,7 @@ function addCommonSDCImportFns(ns) {
                 LForms._valueSetAnswerCache[vsKey] = answers;
                 item.answers = answers;
                 lfData._updateAutocompOptions(item);
+                lfData._resetItemValueWithAnswers(item);
               }
             }).catch(function(error) {
               throw new Error("Unable to load ValueSet "+item.answerValueSet+ " from FHIR server");